+++ app/src/main/java/com/example/playergroup/ui/board/boardDetail/BoardDetailActivity.kt
@@ -1,11 +1,8 @@
 package com.example.playergroup.ui.board.boardDetail
 
 import android.app.Activity
-import android.graphics.Color
 import android.graphics.drawable.GradientDrawable
 import android.os.Bundle
-import android.text.Editable
-import android.text.TextWatcher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.core.content.res.ResourcesCompat
